restructure

+++ Glowworm.cs
@@ -13,9 +13,9 @@ namespace GlowwormSelection
         public int ID { get; set; }
         public int X { get; set; }
         public int Y { get; set; }
-        public float Luciferin { get; set; }
+        public double Luciferin { get; set; }
 
-        public Glowworm(int id, int x, int y, float luciferinStart)
+        public Glowworm(int id, int x, int y, double luciferinStart)
         {
             this.ID = id;
             this.X = x;
